clean deprecated legacy API
