+++ strawberry_django/__init__.py
@@ -1,4 +1,4 @@
-from . import auth, filters, ordering, types
+from . import auth, filters, mutations, ordering, types
 from .fields.field import field
 from .fields.types import (
     DjangoFileType,
@@ -12,10 +12,6 @@ from .fields.types import (
     is_auto,
 )
 from .filters import filter_deprecated as filter
-from .legacy.mutations.auth import AuthMutation
-from .legacy.queries.fields import queries
-from .legacy.registers import TypeRegister
-from .mutations.mutations import mutations
 from .resolvers import django_resolver
 from .type import input, mutation, type
 from .utils import fields
@@ -43,7 +39,4 @@ __all__ = [
     "type",
     "input",
     "mutation",
-    "AuthMutation",
-    "queries",
-    "TypeRegister",
 ]
