--- strawberry_django/legacy/hooks.py
@@ -1,4 +1,0 @@
-from ..hooks import add
-
-
-__all__ = ["add"]
