+++ tests/models.py
@@ -1,10 +1,5 @@
 from django.db import models
 
-from tests.legacy.models import Group, Tag, User
-
-
-__all__ = ["User", "Group", "Tag"]
-
 
 class Fruit(models.Model):
     name = models.CharField(max_length=20)
@@ -27,3 +22,20 @@ class Color(models.Model):
 
 class FruitType(models.Model):
     name = models.CharField(max_length=20)
+
+
+class User(models.Model):
+    name = models.CharField(max_length=50)
+    group = models.ForeignKey(
+        "Group", null=True, related_name="users", on_delete=models.CASCADE
+    )
+    tag = models.OneToOneField("Tag", null=True, on_delete=models.CASCADE)
+
+
+class Group(models.Model):
+    name = models.CharField(max_length=50)
+    tags = models.ManyToManyField("Tag", null=True, related_name="groups")
+
+
+class Tag(models.Model):
+    name = models.CharField(max_length=50)
