+++ tests/queries/test_sync.py
@@ -17,26 +17,3 @@ def test_query(query, user, group, tag):
             },
         }
     ]
-
-
-def test_filters(query, user):
-    result = query(
-        "{ groups { users(filters: [\"name__startswith='us'\","
-        " \"name__contains!='gr'\"]) { name } } }"
-    )
-    assert not result.errors
-    assert result.data["groups"] == [{"users": [{"name": "user"}]}]
-
-    result = query("{ groups { users(filters: [\"name!='user'\"]) { name } } }")
-    assert not result.errors
-    assert result.data["groups"] == [{"users": []}]
-
-
-def test_ordering(query, users):
-    result = query('{ users(orderBy: [ "-name" ]) { name } }')
-    assert not result.errors
-    assert result.data["users"] == [
-        {"name": "user3"},
-        {"name": "user2"},
-        {"name": "user1"},
-    ]
