+++ tests/test_types.py
@@ -1,12 +1,14 @@
 import strawberry_django
+from strawberry_django import auto
 
 from .models import User
 
 
 def test_type_instance():
-    @strawberry_django.type(User, fields=["id", "name"])
+    @strawberry_django.type(User)
     class UserType:
-        pass
+        id: auto
+        name: auto
 
     user = UserType(1, "user")
     assert user.id == 1
@@ -14,9 +16,10 @@ def test_type_instance():
 
 
 def test_input_instance():
-    @strawberry_django.input(User, fields=["id", "name"])
+    @strawberry_django.input(User)
     class InputType:
-        pass
+        id: auto
+        name: auto
 
     user = InputType(1, "user")
     assert user.id == 1
